requirements.txt
django
psutil
GPUtil
django-slik
datetime
py-cpuinfo
socket
